requirements-local.txt
-r requirements.txt
Pillow==11.1.0
pynput==1.7.7
pystray==0.19.5